Refactor project structure and remove testing/demo components
