+++ .gitignore
@@ -6,6 +6,7 @@ yarn-error.log*
 
 # Production builds
 build/
+client/build/
 dist/
 .next/
 out/
