+++ client/src/components/common/Navbar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../../contexts/AuthContext';
 import { LogOut, User } from 'lucide-react';
 
 const Navbar = () => {
