+++ client/src/components/modals/EditEventModal.js
@@ -1,29 +1,17 @@
-import React, { useState, useRef } from 'react';
-import { X, Palette, Upload, Image, File, Trash2 } from 'lucide-react';
-import MediaService from '../services/mediaService';
+import React, { useState, useEffect, useRef } from 'react';
+import { X, Palette, Upload, Image, File, Trash2, Video, Music } from 'lucide-react';
+import MediaService from '../../services/mediaService';
 
-const eventTypes = [
-  { label: 'Memory', value: 'memory', color: '#10B981' }, // green
-  { label: 'Journal', value: 'journal', color: '#3B82F6' }, // blue
-  { label: 'Emotion', value: 'emotion', color: '#EC4899' }, // pink
-  { label: 'Milestone', value: 'milestone', color: '#8B5CF6' }, // purple
-  { label: 'Goal', value: 'goal', color: '#F59E0B' }, // yellow
-];
-
-const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
-  // Find the event type object for initialType
-  const initialTypeObj = eventTypes.find(t => t.value === initialType) || eventTypes[0];
+const EditEventModal = ({ event, onClose, onEdit }) => {
   const [formData, setFormData] = useState({
     title: '',
     description: '',
-    date: new Date().toISOString().slice(0, 16),
+    date: '',
     category: '',
-    type: initialTypeObj.value,
-    color: initialTypeObj.color
+    color: '#3B82F6'
   });
 
   const [showColorPicker, setShowColorPicker] = useState(false);
-  const [colorOverridden, setColorOverridden] = useState(false);
   const [uploadedFiles, setUploadedFiles] = useState([]);
   const [uploading, setUploading] = useState(false);
   const fileInputRef = useRef();
@@ -34,29 +22,25 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
     '#8B5CF6', '#EC4899', '#06B6D4', '#84CC16'
   ];
 
-  // If lockType, do not allow changing type
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    if (name === 'type' && lockType) return;
-    if (name === 'type') {
-      const selectedType = eventTypes.find(t => t.value === value);
-      setFormData(prev => ({
-        ...prev,
-        type: value,
-        color: colorOverridden ? prev.color : (selectedType ? selectedType.color : prev.color)
-      }));
-    } else {
+  useEffect(() => {
+    if (event) {
       setFormData({
-        ...formData,
-        [name]: value
+        title: event.title || '',
+        description: event.description || '',
+        date: new Date(event.date).toISOString().slice(0, 16),
+        category: event.category || '',
+        color: event.color || '#3B82F6'
       });
+      // Initialize uploaded files from existing media
+      setUploadedFiles(event.media || []);
     }
-  };
+  }, [event]);
 
-  const handleColorPick = (color) => {
-    setFormData({ ...formData, color });
-    setColorOverridden(true);
-    setShowColorPicker(false);
+  const handleChange = (e) => {
+    setFormData({
+      ...formData,
+      [e.target.name]: e.target.value
+    });
   };
 
   const handleFileUpload = async (files) => {
@@ -117,12 +101,14 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
         url: file.url,
         path: file.path,
         type: file.type,
-        name: file.originalName
+        name: file.originalName || file.name
       }))
     };
     
-    onAdd(eventData);
+    onEdit(eventData);
   };
+
+  if (!event) return null;
 
   return (
     <div className="modal-overlay">
@@ -131,32 +117,6 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
           <span aria-hidden="true">&times;</span>
         </button>
         <form onSubmit={handleSubmit} className="p-6 space-y-4">
-          {/* Event Type Dropdown or Label */}
-          <div className="form-group">
-            <label htmlFor="type" className="form-label">
-              Event Type *
-            </label>
-            {lockType ? (
-              <div className="input" style={{ background: '#f3f4f6', color: initialTypeObj.color, fontWeight: 600 }}>
-                {initialTypeObj.label}
-              </div>
-            ) : (
-              <select
-                id="type"
-                name="type"
-                value={formData.type}
-                onChange={handleChange}
-                className="input"
-                required
-              >
-                {eventTypes.map((type) => (
-                  <option key={type.value} value={type.value}>{type.label}</option>
-                ))}
-              </select>
-            )}
-          </div>
-
-          {/* Title */}
           <div className="form-group">
             <label htmlFor="title" className="form-label">
               Title *
@@ -218,41 +178,6 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
             />
           </div>
 
-          <div className="form-group">
-            <label className="form-label">Color</label>
-            <div className="flex items-center gap-3">
-              <div
-                className="w-8 h-8 rounded-full border-2 border-gray-300 cursor-pointer"
-                style={{ backgroundColor: formData.color }}
-                onClick={() => setShowColorPicker(!showColorPicker)}
-              />
-              <button
-                type="button"
-                onClick={() => setShowColorPicker(!showColorPicker)}
-                className="flex items-center gap-2 text-gray-600 hover:text-gray-800"
-              >
-                <Palette size={16} />
-                Choose Color
-              </button>
-            </div>
-            {showColorPicker && (
-              <div className="mt-3 flex gap-2 flex-wrap">
-                {colorOptions.map((color) => (
-                  <button
-                    key={color}
-                    type="button"
-                    className="w-8 h-8 rounded-full border-2 border-gray-300 hover:border-gray-400 transition-colors"
-                    style={{ backgroundColor: color }}
-                    onClick={() => handleColorPick(color)}
-                  />
-                ))}
-              </div>
-            )}
-            <div className="text-xs text-gray-400 mt-1">
-              Default color is based on event type. You can override it.
-            </div>
-          </div>
-
           {/* File Upload Section */}
           <div className="form-group">
             <label className="form-label">Media Files</label>
@@ -286,12 +211,18 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
                 <div className="grid grid-cols-2 gap-2">
                   {uploadedFiles.map((file, index) => (
                     <div key={index} className="flex items-center gap-2 p-2 bg-gray-50 rounded">
-                      {file.type.startsWith('image/') ? (
+                      {file.type && file.type.startsWith('image/') ? (
                         <Image size={16} className="text-blue-500" />
+                      ) : file.type && file.type.startsWith('video/') ? (
+                        <Video size={16} className="text-purple-500" />
+                      ) : file.type && file.type.startsWith('audio/') ? (
+                        <Music size={16} className="text-green-500" />
                       ) : (
                         <File size={16} className="text-gray-500" />
                       )}
-                      <span className="text-xs truncate flex-1">{file.originalName}</span>
+                      <span className="text-xs truncate flex-1">
+                        {file.originalName || file.name || 'File'}
+                      </span>
                       <button
                         type="button"
                         onClick={() => removeFile(index)}
@@ -302,6 +233,42 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
                     </div>
                   ))}
                 </div>
+              </div>
+            )}
+          </div>
+
+          <div className="form-group">
+            <label className="form-label">Color</label>
+            <div className="flex items-center gap-3">
+              <div
+                className="w-8 h-8 rounded-full border-2 border-gray-300 cursor-pointer"
+                style={{ backgroundColor: formData.color }}
+                onClick={() => setShowColorPicker(!showColorPicker)}
+              />
+              <button
+                type="button"
+                onClick={() => setShowColorPicker(!showColorPicker)}
+                className="flex items-center gap-2 text-gray-600 hover:text-gray-800"
+              >
+                <Palette size={16} />
+                Choose Color
+              </button>
+            </div>
+            
+            {showColorPicker && (
+              <div className="mt-3 flex gap-2 flex-wrap">
+                {colorOptions.map((color) => (
+                  <button
+                    key={color}
+                    type="button"
+                    className="w-8 h-8 rounded-full border-2 border-gray-300 hover:border-gray-400 transition-colors"
+                    style={{ backgroundColor: color }}
+                    onClick={() => {
+                      setFormData({ ...formData, color });
+                      setShowColorPicker(false);
+                    }}
+                  />
+                ))}
               </div>
             )}
           </div>
@@ -318,7 +285,7 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
               type="submit"
               className="btn btn-primary flex-1"
             >
-              Add Event
+              Update Event
             </button>
           </div>
         </form>
@@ -327,4 +294,4 @@ const AddEventModal = ({ onClose, onAdd, initialType, lockType }) => {
   );
 };
 
-export default AddEventModal; 
+export default EditEventModal;
