+++ client/src/components/modals/FocusedEventModal.js
