--- client/src/lib/utils.js
@@ -1,4 +1,0 @@
-// Utility function for conditionally joining classNames
-export function cn(...inputs) {
-  return inputs.filter(Boolean).join(' ');
-}
